requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic[email]>=2.9.0
requests>=2.32.0
flask>=3.1.0
gunicorn>=21.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.2.0,<4.0.0
python-multipart>=0.0.9
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
